Prevent DefaultGroup deletion or modification. Prevent to add a group which's name is "DefaultGroup". Create a default group object. Resolving issue #34.

+++ app/models/Group.scala
@@ -15,6 +15,23 @@ import play.api.db._
 // Defining a case class
 case class Group(id: Long, name: String)
 
+/**
+ * Default group definition
+ */
+object DefaultGroup {
+  /**
+   * Default group identifiers
+   */
+  private val defaultGroupId = 1;
+  private val defaultGroupName = "DefaultGroup";
+
+  /**
+   * Default group value
+   */
+  val defaultGroupObject = new Group(defaultGroupId,defaultGroupName);
+
+}
+
 object Group {
 
   /**
@@ -23,6 +40,8 @@ object Group {
   private val keyCacheAllOptions = "group-options"
   private val keyCacheById = "group-by-id"
   private val keyCacheByName = "group-by-name"
+
+
 
   /**
    * SQL anorm row parser. This operation indicate how to parse a sql row.
@@ -103,12 +122,15 @@ object Group {
    * @param group The group to persist.
    */
   def insert(group: Group) = {
-    // Clear the cache in order to ???
-    clearCache
-    // Insert the new group
-    DB.withConnection {
-      implicit connection =>
-        SQL( """	insert into groups values (null, {name})""").on('name -> group.name).executeUpdate()
+    // A group which looks like the default group cannot be inserted
+    if(group.id != DefaultGroup.defaultGroupObject.id && group.name != DefaultGroup.defaultGroupObject.name){
+      // Clear the cache in order to ???
+      clearCache
+      // Insert the new group
+      DB.withConnection {
+        implicit connection =>
+          SQL( """	insert into groups values (null, {name})""").on('name -> group.name).executeUpdate()
+      }
     }
   }
 
@@ -118,14 +140,17 @@ object Group {
    * @param group The group group.
    */
   def update(group: Group) = {
-    clearCache
-    Cache.remove(keyCacheById + group.id)
-    DB.withConnection {
-      implicit connection =>
-        SQL( """update groups set name = {name} where id = {id}""").on(
-          'id -> group.id,
-          'name -> group.name
-        ).executeUpdate()
+    //The default group cannot be modified
+    if(group.id != DefaultGroup.defaultGroupObject.id){
+      clearCache
+      Cache.remove(keyCacheById + group.id)
+      DB.withConnection {
+        implicit connection =>
+          SQL( """update groups set name = {name} where id = {id}""").on(
+            'id -> group.id,
+            'name -> group.name
+          ).executeUpdate()
+      }
     }
   }
 
@@ -135,11 +160,15 @@ object Group {
    * @param group group to delete
    */
   def delete(group: Group) = {
-    clearCache
-    Cache.remove(keyCacheById + group.id)
-    DB.withConnection {
-      implicit connection =>
-        SQL("delete from groups where id = {id}").on('id -> group.id).executeUpdate()
+
+    // The default group cannot be delete
+    if(group.id != DefaultGroup.defaultGroupObject.id){
+      clearCache
+      Cache.remove(keyCacheById + group.id)
+      DB.withConnection {
+        implicit connection =>
+          SQL("delete from groups where id = {id}").on('id -> group.id).executeUpdate()
+      }
     }
   }
